requirements.txt
streamlit==1.37.1
pandas==2.2.2
openpyxl==3.1.5
google-auth==2.33.0
gspread==6.1.2
python-dateutil==2.9.0.post0